pyproject.toml
[project]
name = "uv-tui"
version = "0.1.0"
description = "A TUI for managing Python virtual environments with uv."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rich>=14.1.0",
    "rich-color-ext>=0.1.4",
    "rich-gradient>=0.3.6",
    "textual-dev>=1.7.0",
]
